# Issue

The no warning settings have become outdated.

```
cc1plus: note: unrecognized command-line option ‘-Wno-shorten-64-to-32’ may have been intended to silence earlier diagnostics
cc1plus: note: unrecognized command-line option ‘-Wno-unused-member-function’ may have been intended to silence earlier diagnostics
cc1plus: note: unrecognized command-line option ‘-Wno-implicit-float-conversion’ may have been intended to silence earlier diagnostics
cc1plus: note: unrecognized command-line option ‘-Wno-implicit-int-conversion’ may have been intended to silence earlier diagnostics
```

<!-- BENCHMARK_RESULTS_START -->
## Benchmark Results
| Benchmark | Base | PR |
|-----------|------|----|
| alias | aliased u32: 1088.9<br/>plain u32: 1094.37<br/>aliased double: 954.818<br/>plain double: 952.287 | aliased u32: 1080.74<br/>plain u32: 1080.86<br/>aliased double: 940.798<br/>plain double: 948.737 |
| json-render | String: 8.39944ms<br/>Stringstream: 11.7702ms<br/>Vector: 7.59087ms | String: 8.47566ms<br/>Stringstream: 11.8241ms<br/>Vector: 7.61138ms |
| match_ch | Default radius: <br/>4.48361ms/req at 82 coordinate<br/>0.0546781ms/coordinate<br/>Radius 5m: <br/>4.46375ms/req at 82 coordinate<br/>0.054436ms/coordinate<br/>Radius 10m: <br/>15.274ms/req at 82 coordinate<br/>0.186268ms/coordinate<br/>Radius 15m: <br/>37.176ms/req at 82 coordinate<br/>0.453366ms/coordinate<br/>Radius 30m: <br/>315.754ms/req at 82 coordinate<br/>3.85066ms/coordinate | Default radius: <br/>4.48168ms/req at 82 coordinate<br/>0.0546547ms/coordinate<br/>Radius 5m: <br/>4.44683ms/req at 82 coordinate<br/>0.0542297ms/coordinate<br/>Radius 10m: <br/>15.1663ms/req at 82 coordinate<br/>0.184954ms/coordinate<br/>Radius 15m: <br/>37.1405ms/req at 82 coordinate<br/>0.452933ms/coordinate<br/>Radius 30m: <br/>313.781ms/req at 82 coordinate<br/>3.8266ms/coordinate |
| match_mld | Default radius: <br/>3.45143ms/req at 82 coordinate<br/>0.0420907ms/coordinate<br/>Radius 5m: <br/>3.43708ms/req at 82 coordinate<br/>0.0419156ms/coordinate<br/>Radius 10m: <br/>12.4145ms/req at 82 coordinate<br/>0.151396ms/coordinate<br/>Radius 15m: <br/>31.8435ms/req at 82 coordinate<br/>0.388335ms/coordinate<br/>Radius 30m: <br/>355.008ms/req at 82 coordinate<br/>4.32936ms/coordinate | Default radius: <br/>3.37035ms/req at 82 coordinate<br/>0.0411019ms/coordinate<br/>Radius 5m: <br/>3.35458ms/req at 82 coordinate<br/>0.0409095ms/coordinate<br/>Radius 10m: <br/>12.3371ms/req at 82 coordinate<br/>0.150452ms/coordinate<br/>Radius 15m: <br/>31.8716ms/req at 82 coordinate<br/>0.388678ms/coordinate<br/>Radius 30m: <br/>352.352ms/req at 82 coordinate<br/>4.29697ms/coordinate |
| packedvector | random write:<br/>std::vector 10465.7 ms<br/>util::packed_vector 82037.9 ms<br/>slowdown: 7.83874<br/>random read:<br/>std::vector 8428.56 ms<br/>util::packed_vector 32889.1 ms<br/>slowdown: 3.9021 | random write:<br/>std::vector 9702.51 ms<br/>util::packed_vector 73755.9 ms<br/>slowdown: 7.60173<br/>random read:<br/>std::vector 8368.22 ms<br/>util::packed_vector 30225.7 ms<br/>slowdown: 3.61196 |
| route_ch | 1000 routes, 3 coordinates, no alternatives, overview=full, steps=true<br/>598.507ms<br/>0.598507ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=full, steps=true<br/>376.315ms<br/>0.376315ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=full, steps=true<br/>757.937ms<br/>0.757937ms/req<br/>1000 routes, 3 coordinates, no alternatives, overview=false, steps=false<br/>152.07ms<br/>0.15207ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=false, steps=false<br/>96.7528ms<br/>0.0967528ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false<br/>134.199ms<br/>0.134199ms/req<br/>1000 routes, 3 coordinates, no alternatives, overview=false, steps=false, radius=750<br/>150.686ms<br/>0.150686ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=false, steps=false, radius=750<br/>96.5346ms<br/>0.0965346ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false, radius=750<br/>133.009ms<br/>0.133009ms/req | 1000 routes, 3 coordinates, no alternatives, overview=full, steps=true<br/>596.313ms<br/>0.596313ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=full, steps=true<br/>375.496ms<br/>0.375496ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=full, steps=true<br/>757.985ms<br/>0.757985ms/req<br/>1000 routes, 3 coordinates, no alternatives, overview=false, steps=false<br/>153.084ms<br/>0.153084ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=false, steps=false<br/>98.4702ms<br/>0.0984702ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false<br/>133.543ms<br/>0.133543ms/req<br/>1000 routes, 3 coordinates, no alternatives, overview=false, steps=false, radius=750<br/>151.211ms<br/>0.151211ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=false, steps=false, radius=750<br/>97.8749ms<br/>0.0978749ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false, radius=750<br/>133.784ms<br/>0.133784ms/req |
| route_mld | 1000 routes, 3 coordinates, no alternatives, overview=full, steps=true<br/>729.415ms<br/>0.729415ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=full, steps=true<br/>461.985ms<br/>0.461985ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=full, steps=true<br/>939.481ms<br/>0.939481ms/req<br/>1000 routes, 3 coordinates, no alternatives, overview=false, steps=false<br/>275.042ms<br/>0.275042ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=false, steps=false<br/>160.646ms<br/>0.160646ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false<br/>286.759ms<br/>0.286759ms/req<br/>1000 routes, 3 coordinates, no alternatives, overview=false, steps=false, radius=750<br/>271.89ms<br/>0.27189ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=false, steps=false, radius=750<br/>160.312ms<br/>0.160312ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false, radius=750<br/>286.314ms<br/>0.286314ms/req | 1000 routes, 3 coordinates, no alternatives, overview=full, steps=true<br/>721.265ms<br/>0.721265ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=full, steps=true<br/>459.617ms<br/>0.459617ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=full, steps=true<br/>935.905ms<br/>0.935905ms/req<br/>1000 routes, 3 coordinates, no alternatives, overview=false, steps=false<br/>266.621ms<br/>0.266621ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=false, steps=false<br/>161.579ms<br/>0.161579ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false<br/>286.558ms<br/>0.286558ms/req<br/>1000 routes, 3 coordinates, no alternatives, overview=false, steps=false, radius=750<br/>266.104ms<br/>0.266104ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=false, steps=false, radius=750<br/>160.24ms<br/>0.16024ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false, radius=750<br/>288.362ms<br/>0.288362ms/req |
| rtree | 1 result:<br/>206.655ms ->  0.0206655 ms/query<br/>10 results:<br/>241.654ms ->  0.0241654 ms/query | 1 result:<br/>206.761ms ->  0.0206761 ms/query<br/>10 results:<br/>242.589ms ->  0.0242589 ms/query |
<!-- BENCHMARK_RESULTS_END -->